Implement bilayers
